# server/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ZweckOS API",
    "description": "Internal accounting and operations API. Authenticated routes except `/auth/login`, `/auth/register`, `/auth/forgot-password`, `/auth/reset-password`, `/health`, `/docs`, `/openapi.json`.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/api", "description": "Relative to API origin" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginBody": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } }
      }
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Sign in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Request password reset email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": { "email": { "type": "string", "format": "email" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic success (does not reveal whether email exists)"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Set new password with token from email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token", "password"],
                "properties": {
                  "token": { "type": "string" },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password updated" },
          "400": { "description": "Invalid or expired token" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": ["transactions"],
        "summary": "List transactions",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["transactions"],
        "summary": "Create transaction",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/accounts": {
      "get": {
        "tags": ["accounts"],
        "summary": "Chart of accounts / balances",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/directors": {
      "get": {
        "tags": ["directors"],
        "summary": "List directors",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/portfolio": {
      "get": {
        "tags": ["portfolio"],
        "summary": "Portfolio summary",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "List projects",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Create project",
        "security": [{ "bearerAuth": [] }],
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "List users (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/audit": {
      "get": {
        "tags": ["audit"],
        "summary": "Audit log (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/settings": {
      "get": {
        "tags": ["settings"],
        "summary": "Settings / monitoring summary (authenticated)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Version, uptime, monitoring flags, rate limits" } }
      }
    }
  }
}